refactor: separate Milestones from ProfileGallery
